+++ client/src/App.js
@@ -17,7 +17,7 @@ const DefaultLayout = React.lazy(() => import('./layout/DefaultLayout'))
 const Login = React.lazy(() => import('./views/pages/login/Login'))
 const RegisterEntry = React.lazy(() => import('./views/pages/register/RegisterEntry'))
 const Register = React.lazy(() => import('./views/pages/register/Register'))
-const RegisterFB = React.lazy(() => import('./views/pages/registerFB/RegisterFB'))
+const RegisterFB = React.lazy(() => import('./views/pages/register/RegisterFB'))
 const Forget = React.lazy(() => import('./views/pages/forget/Forget'))
 const Page404 = React.lazy(() => import('./views/pages/page404/Page404'))
 const Page500 = React.lazy(() => import('./views/pages/page500/Page500'))
